Add xbuf for response buffer
Previous version used String which did not work with binary response
data.  Also fairly inefficient.  xbuf is a buffer composed of dynamic
segments that delete from the head as data is removed and allocate to
the tail as data is added.  Segment size is compiled in with a #define,
currently 64 bytes - so 12% overhead.
needed this to support buffer search to parse HTTP headers and chunk
headers.  Probably reinventing the wheel, but it took less time than
researching all the undocumented ?buf out there.